+++ changelog.md
@@ -7,6 +7,7 @@
 - load native taskbar pins, for new users.
 - show task switcher on the monitor where the cursor is positioned.
 - add option to filter task switcher apps by the active monitor.
+- customizable twm float size
 
 ### fix